requirements.txt
# CypherD Wallet Backend - SQLite Version
# BULLETPROOF dependencies for Web3 wallet operations

# Flask Framework
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
Flask-SQLAlchemy==3.0.5

# Database
SQLAlchemy==2.0.23

# Environment Management
python-dotenv==1.0.0

# HTTP Requests
requests==2.31.0

# Cryptography
cryptography==41.0.4
eth-account==0.8.0
mnemonic==0.20
bcrypt==4.0.1

# Email
email-validator==2.0.0

# Development Dependencies (optional)
pytest==7.4.3
pytest-flask==1.3.0
